requirements.txt
python-dotenv==1.0.1

deltalake==0.25.4
pandas==2.2.3

solders==0.26.0
solana==0.36.6

aiohttp==3.11.13
starlette==0.46.1

confluent-kafka==2.8.2
pydantic==2.10.6
protobuf==6.30.1

# ZMQ socket BTC connection
pyzmq==26.3.0
